Continue updating tests
