+++ tests/unit/api/test_variables.py
@@ -19,8 +19,6 @@ def test_variables_uri(flask_app, variable, expected):
     assert variables.uri(variable) == expected
 
 
-@pytest.mark.xfail(reason="Needs a version of PyCDS that has not yet been released")
-# TODO: Update test to include tags attribute
 def test_variable_collection(everything_session, tst_networks, tst_variables):
     vars = sorted(variables.collection(), key=lambda r: r["id"])
     assert vars == [
@@ -35,14 +33,13 @@ def test_variable_collection(everything_session, tst_networks, tst_variables):
             "unit": var.unit,
             "precision": var.precision,
             "network_uri": networks.uri(var.network),
+            "tags": ['observation']
         }
         for var in tst_variables
         if var.network == tst_networks[0]
     ]
 
 
-@pytest.mark.xfail(reason="Needs a version of PyCDS that has not yet been released")
-# TODO: Update test to include tags attribute
 def test_variable_item(everything_session, tst_networks, tst_variables):
     for var in tst_variables:
         if var.network == tst_networks[0]:
@@ -58,4 +55,5 @@ def test_variable_item(everything_session, tst_networks, tst_variables):
                 "unit": var.unit,
                 "precision": var.precision,
                 "network_uri": networks.uri(var.network),
+                "tags": ['observation']
             }
